+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * `[[JavaStack]]` for a Java exception (was previously a regular property)
     * `[[Message]]` for a Java exception (was previously a regular property)
 * Don't return any Java array properties if only accessors are requested.
+* Merge scripts based on contents hash regardless of recompilation/original order (issue #64)
 
 ## 0.5.5 (2017-08-24)